MB-27661 [8/n]: Fix MSVC warnings - C4244
This patch addresses the following generated warnings:

C4244 - conversion' conversion from 'type1' to 'type2', possible
	loss of data. An integer type is converted to a smaller integer
	type.

Reviewed-on: http://review.couchbase.org/89980
Tested-by: Build Bot <build@couchbase.com>

+++ auditd/generator/auditevent_generator.h
@@ -18,6 +18,7 @@
 #ifndef AUDIT_DESCRIPTORS_H
 #define AUDIT_DESCRIPTORS_H
 
+#include <gsl/gsl>
 #include <string>
 
 #define eventid_modulus 4096
@@ -128,7 +129,7 @@ public:
         cJSON *fname = getMandatoryObject(data, "file", cJSON_String);
         cJSON *hfile = getOptionalObject(data, "header", cJSON_String);
 
-        start = sid->valueint;
+        start = gsl::narrow<int>(sid->valueint);
         file.assign(srcRoot);
         file.append("/");
         file.append(fname->valuestring);
